Initalized and styled UI layout for create-account page
Add home helper module and view template

This commit adds a new home helper module and view template. The  module is created in the  directory, and the  view template is added to the  directory. These changes are necessary to implement the home page functionality.

API is built out and functional. App is ready to go
Updating test cases now